* Cleanups all over the code.
* interrogator.h: Now we have this new file that defines the two classes
TransferInterrogator and GroupInterrogator. If a class wants to interrogate
a transfer or a group object it should reimplement them.

* group.{h,cpp}: The progressed size, totalSize and percent properties are
not saved anymore to file. Now we update the group properties every time
we see that a transfer that belongs to the group has changed. Each group
change is notified to all the views.

* mainview.cpp: fixed crash when executing actions on a group item.


git-svn-id: svn+ssh://svn.kde.org/home/kde/branches/make_kget_cool/kdenetwork/kget@360582 283d02a7-25f6-0310-bc7c-ecb5cbfe19da