MultisegKIO now use KIO::Get for the segments and KIO::file for the local file. this should fix many problems with http redireccions and so. ftp also works now.
I make a segmentfactory class that handle the segments and olso can be used in the metalink plugin

now we pass a QList<KUrl> to the MultiSegfile_copy funtion. i'm planing to make a search funtion to return an KUrl List. any help or idea is welcome :)

CCMAIL: kget@kde.org



git-svn-id: svn+ssh://svn.kde.org/home/kde/branches/work/make_kget_cool/kget@604618 283d02a7-25f6-0310-bc7c-ecb5cbfe19da